Reorganized base resource files
