+++ src/resources/base/base.js
@@ -1,6 +1,6 @@
-const ENV = require('../environment')();
+const ENV = require('../../environment')();
 const mongodb = require('mongodb');
-const dateService = require('../services/date-service');
+const dateService = require('../../services/date/date');
 
 const _public = {};
